backend/pyproject.toml
[project]
name = "polidex"
version = "0.1.0"
description = "RAG Admin System for managing knowledge base documents"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.1",
    "pymupdf>=1.23.8",
    "python-docx>=1.1.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "slowapi>=0.1.9",
    "argon2-cffi>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
